third_party/WebKit/LayoutTests/http/tests/activedomobject/media-expected.txt

Tests that reparenting media elements also reparents ActiveDOMObject.


Before Reparenting
PASS: internals.pausableObjectCount(document) should be '2' and is.
PASS: internals.pausableObjectCount(iframe) should be '2' and is.
After Reparenting
PASS: internals.pausableObjectCount(document) should be '3' and is.
PASS: internals.pausableObjectCount(iframe) should be '1' and is.
